refactor(api/common): use built-in ExceptionHandlerMiddleware
