+++ modules/back-end/tests/Application.IntegrationTests/Snapshots/IdentityControllerTests.LoginByPassword_RequestValidation.verified.txt
@@ -2,11 +2,14 @@
   Version: 1.1,
   Status: 400 Bad Request,
   Headers: {
-    api-supported-versions: 1.0
+    api-supported-versions: 1.0,
+    Cache-Control: no-store, no-cache,
+    Pragma: no-cache
   },
   Content: {
     Headers: {
-      Content-Type: application/json; charset=utf-8
+      Content-Type: application/json; charset=utf-8,
+      Expires: -1
     },
     Value: {
       data: null,
